+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased changes
 
+- Clean up and document LPPaver.Decide.Util
 - Split up single Type.hs module into smaller and better encapsulated modules
 - Add constraintRightSide to LPPaver.Constraint.Util
 - Document LPPaver.Constraint.Type and LPPaver.Constraint.Util